+++ packages/ui/src/index.ts
@@ -42,3 +42,4 @@ export * from './components/SectionLabel';
 export * from './components/Toast';
 export * from './components/Combobox';
 export * from './components/TableFilters';
+export * from './components/ChipPicker';